requirements.txt
# Copyright (c) 2017-2019 The University of Manchester
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.

six
enum34
SpiNNUtilities >= 1!4.0.1, < 1!5.0.0
SpiNNMachine >= 1!4.0.1, < 1!5.0.0
SpiNNStorageHandlers >= 1!4.0.1, < 1!5.0.0
SpiNNMan >= 1!4.0.1, < 1!5.0.0
SpiNNaker_PACMAN >= 1!4.0.1, < 1!5.0.0
SpiNNaker_DataSpecification >= 1!4.0.1, < 1!5.0.0
spalloc >= 1.0.1, < 2.0.0
requests >= 2.4.1
scipy >= 0.16.0
numpy
futures; python_version == "2.7"
# spinnaker_tools
